# KAFKA-10074: Improve performance of `matchingAcls`