@ijuma 2020-06-01
The issue affecting the Scala 2.12 build is unrelated. Looks like Gradle exited while running Streams tests.